receive refactor
